% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_splitting.R
\name{create_data_partition}
\alias{create_data_partition}
\alias{create_kfolds}
\alias{create_bootstrap}
\title{Data splitting functions for uplift.}
\usage{
create_data_partition(y, trt = NULL, p = 0.5, times = 1, groups = 5,
  replace = FALSE)

create_kfolds(y, trt = NULL, k = 10, times = 1, groups = 5)

create_bootstrap(y, times = 10)
}
\arguments{
\item{y}{An atomic vector.}

\item{trt}{An optional treatment variable.}

\item{p}{The proportion of training observations.}

\item{times}{The number of partitions to create.}

\item{groups}{For numeric y, the number of breaks in the quantiles.}

\item{replace}{Should sampling be done with replacement?}

\item{k}{The number of folds.}
}
\value{
\code{create_data_partition} and \code{create_bootstrap} return a
 matrix of row position integers corresponding to the training set and to the
 bootstrap sample, respectively. \code{create_kfolds} returns a matrix with
 the row integers corresponding to the folds.
}
\description{
\code{create_data_partition} creates one or more random data partitions into
training and test sets. \code{create_kfolds} splits the data into k-folds (or
groups) with approximatly the same number of observations.
\code{create_bootstrap} draws bootstrap samples.
}
\details{
If \code{y} is a factor, sampling is done within the levels of \code{y} in an
attempt to balance the class distributions between the partitions. If \code{y}
is numeric, groups are first created based on the quantiles of its
distribution and then sampling is done within these groups.

If \code{trt} is supplied, the data partitions are stratified by the treatment
variable.

Notice that in addition to \code{create_bootstrap}, bootstrap samples can also
by created using \code{create_data_partition} with \code{p = 1} and
\code{replace = TRUE}.
}
\examples{

set.seed(545)
r <- factor(sample(c(0,1), 1000, replace = TRUE))
t <- factor(sample(c(0,1), 1000, replace = TRUE))
df <- data.frame(r, t)
trainIndex <- create_data_partition(df$r, df$t)
dfTrain <- df[trainIndex, ]
dfTest <- df[-trainIndex, ]
table(df$r, df$t)
table(dfTrain$r, dfTrain$t)
table(dfTest$r, dfTest$t)
# Create k-folds
head(create_kfolds(r, t, times = 5))

# Create 10 bootstrap samples
set.seed(1)
x <- rnorm(100)
xb <- create_bootstrap(x, times = 10)
}
\author{
Leo Guelman \email{leo.guelman@gmail.com}
}
