% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_uplift_formula.R
\name{create_uplift_formula}
\alias{create_uplift_formula}
\title{Create an uplift formula.}
\usage{
create_uplift_formula(x, y, trt, env = parent.frame())
}
\arguments{
\item{x}{A character vector of predictor names.}

\item{y}{A character vector of the response name.}

\item{trt}{A character vector of the treatment name.}

\item{env}{The environment associated with the result, the calling environment
by default.}
}
\value{
A formula object.
}
\description{
\code{create_uplift_formula} is a helper function to create a formula object
as required by most functions in the uplift package.
}
\examples{
set.seed(1)
df <- sim_uplift(n = 100, p = 20, response = "binary")
f <- create_uplift_formula(names(df)[-c(1:3)], "y", "T")
class(f)
environment(f) # the callling environment
}
\author{
Leo Guelman \email{leo.guelman@gmail.com}
}
