% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_balance.R
\name{ggplot.inspect_balance}
\alias{ggplot.inspect_balance}
\title{Plot a `inspect_balance' object.}
\usage{
\method{ggplot}{inspect_balance}(x, i.var = NULL, n.type = "boxplot",
  f.type = "percent", legend.position = NULL, nrows = NULL,
  ncols = NULL)
}
\arguments{
\item{x}{An object of class \code{inspect_balance}.}

\item{i.var}{A numeric vector of indices of the variables to plot. The
variables should be indexed in the same order that they appear in the
initial inspect_balance formula. The default is to plot all variables.}

\item{n.type}{The type of plot for numeric variables. Boxplots are generated
by default. Alternatively, use \code{n.type = "density"} for density plots,
or \code{n.type = "qqplot"} for qq-plots.}

\item{f.type}{The type of plot for categorical variables. 100-percent stacked
columns are generated by default. The alternative option is \code{"counts"},
which show stacked columns of counts.}

\item{legend.position}{The position of legends \code{("left", "right",
"bottom", "top")}. The default is \code{"right"}.}

\item{nrows}{Number of rows for plots.}

\item{ncols}{Number of columns for plots.}
}
\description{
ggplot method for class \code{inspect_balance}.
}
\examples{
set.seed(343)
df <- sim_uplift(n = 200, p = 50, response = "binary")
df$T <- ifelse(df$T == 1, 1, 0)
ib <- inspect_balance(T~ X1 + X2 + X3, data = df, method ="pdev", nPerm = 500)
ggplot(ib)
ggplot(ib, n.type = "density")
ggplot(ib, i.var = c(1,2), ncols =2, n.type ="density")
}
\author{
Leo Guelman \email{leo.guelman@gmail.com}
}
