% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_performance.R
\name{ggplot.inspect_performance}
\alias{ggplot.inspect_performance}
\title{Qini and Calibration plots from a \code{inspect_performance} object.}
\usage{
\method{ggplot}{inspect_performance}(x, type = "qini",
  legend.position = "top", fillCol = "white", facets = FALSE,
  pointSize = FALSE, diagCol = "grey", xlim = NULL, ylim = NULL,
  xBreaks = NULL, yBreaks = NULL, xlab = NULL, ylab = NULL,
  title = NULL)
}
\arguments{
\item{x}{An object of class \code{"inspect_performance"}.}

\item{type}{The type of plot. Possible values are \code{qini} for Qini curves
(default) and \code{calib} for calibration plots.}

\item{legend.position}{The position of legends ("left", "right", "bottom",
"top").}

\item{fillCol}{Panel's background color.}

\item{facets}{Lay out panels in a grid?}

\item{pointSize}{For \code{type = "calib"}, make size of points in plot
proportional to the number of observations?}

\item{diagCol}{For \code{type = "calib"}, the color of the diagonal line.}

\item{xlim, ylim}{Numeric vectors of length 2, giving the x and y coordinates
ranges.}

\item{xBreaks, yBreaks}{Points at which x, y gridlines appear.}

\item{xlab, ylab}{Title for the x, y axes.}

\item{title}{The main title for the plot.}
}
\description{
ggplot method for class `inspect_performance'
}
\examples{

set.seed(324)
df_train <- sim_uplift(p = 30, response = "binary")
df_test <- sim_uplift(n = 10000, p = 30, response = "binary")
fit_t1 <- glm(as.formula(paste('y ~', paste('X', 1:30, sep = '', collapse = "+"))),
             family = "binomial", data = df_train, subset = T==1)
fit_t0 <- glm(as.formula(paste('y ~', paste('X', 1:30, sep = '', collapse = "+"))),
             family = "binomial", data = df_train, subset = T==-1)
uplift_score <- predict(fit_t1, df_test, type = "response") -
               predict(fit_t0, df_test, type = "response")
df_test$uplift_score <- uplift_score
res <- inspect_performance(y ~ uplift_score + trueUplift + trt(T),
                          data = df_test, qini = TRUE)
res
summary(res)
ggplot(res)
ggplot(res, type = "calib", pointSize = TRUE)
}
\author{
Leo Guelman \email{leo.guelman@gmail.com}
}
