% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niv.R
\name{ggplot.niv}
\alias{ggplot.niv}
\title{Net weight of evidence plots from a `niv' object.}
\usage{
\method{ggplot}{niv}(x, i.var = NULL, same.limits = FALSE, col = NULL,
  xlab = NULL, ylab = NULL, title = NULL, nrows = NULL, ncols = NULL)
}
\arguments{
\item{x}{An object of class \code{niv}.}

\item{i.var}{A numeric vector of indices of the variables to plot. The
variables should be indexed in the same order that they appear in the
initial niv formula. The default is to plot all variables.}

\item{same.limits}{Should limits for the y-axis be the same in all plots?}

\item{col}{The fill colour of the bars.}

\item{xlab}{A character string giving the title for the x axis (the length
should be the same as the number of plots).}

\item{ylab}{A character string of length 1 giving the title for the y axis.}

\item{title}{The main title for the plot (the length should be the same as the
number of plots).}

\item{nrows}{Number of rows for plots.}

\item{ncols}{Number of columns for plots.}
}
\description{
ggplot method for class \code{niv}.
}
\examples{
set.seed(1)
df <- sim_uplift(n = 1000, p = 20, response = "binary")
netInf <- niv(y ~ trt(T) + X1 + X2 + X9 + X10, data = df,
             B=10, parallel = FALSE, digitsB = 1)
ggplot(netInf)
}
\author{
Leo Guelman \email{leo.guelman@gmail.com}
}
