% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niv.R
\name{niv}
\alias{niv}
\alias{niv.formula}
\alias{print.niv}
\alias{summary.niv}
\title{Net Information Value.}
\usage{
\method{niv}{formula}(formula, data, subset, na.action = na.pass,
  nBins = 10, continuous = 4, B = 10, woeAdj = 0.5, parallel = TRUE,
  nCore = NULL, digitsB = NULL, classLevel = NULL, treatLevel = NULL)

\method{print}{niv}(x, ...)

\method{summary}{niv}(object, ...)
}
\arguments{
\item{formula}{A model formula of the form y ~ x1 + ....+ xn + trt(), where
the left-hand side corresponds to the observed response, the right-hand side
corresponds to the predictors, and 'trt' is the special expression to mark
the treatment term. If the treatment term is not a factor, it is converted to one.
\code{niv} only handles response variables of class factor.}

\item{data}{A data frame in which to interpret the variables named in the
formula.}

\item{subset}{Expression indicating which subset of the rows of data should be
included. All observations are included by default.}

\item{na.action}{A missing-data filter function. Defaults to \code{na.pass}.}

\item{nBins}{The number of bins created from numeric predictors. The bins are
created based on sample quantiles, with a default value of 10 (deciles).}

\item{continuous}{Specifies the threshold for when bins should be created from
numeric predictors. If there are less or equal than n (i.e.,
\code{continuous = n}) unique values in the numeric predictor, it is
coverted to a factor without binning. The default is \code{continuous = 4}.}

\item{B}{The number of bootstraps.}

\item{woeAdj}{The adjustment factor used to avoid an undefined WOE. The value
should be between [0, 1]. By default \code{woeAdj = 0.5}. See details.}

\item{parallel}{If \code{TRUE}, computations are performed in parallel,
otherwise they are done sequentially.}

\item{nCore}{The number of cores used. Default is: number of available
cores-1.}

\item{digitsB}{Number of digits used in formatting the breaks in numeric
predictors.}

\item{classLevel}{A character string for the class of interest. Defaults to
the last level of the factor.}

\item{treatLevel}{A character string for the treatment level of interest.
Defaults to the last level of the treatment factor.}

\item{x}{A \code{niv} object.}

\item{object}{A \code{niv} object.}
}
\value{
An object of class \code{niv}, which is a list with the following
 components (among others passed to the S3 methods): \itemize{ \item
 \code{nwoeData} A list of data frames, one for each variable. The columns
 represent: \itemize{ \item \code{y00} the number of non-event records
 (response != \code{classLevel}) in the control group (treatment !=
 \code{treatLevel}). \item \code{y10} the number of event records (response
 == \code{classLevel}) in the control group (treatment != \code{treatLevel}).
 \item \code{y01} the number of non-event records in the treatment group
 (treatment == \code{treatLevel}). \item \code{y11} the number of event
 records in the treatment group. \item \code{py00} proportion of non-event
 records in the control group. \item \code{py10} proportion of event records
 in the control group. \item \code{py01} proportion of non-event records in
 the treatment group. \item \code{py11} proportion of event records in the
 treatment group. \item \code{woe0} the control group weight-of-evidence.
 \item \code{woe1} the treatment group weight-of-evidence. \item \code{nwoe}
 the net weight-of-evidence. \item \code{niv} the net information value. }
 The values above are computed based on the entire data. \item \code{nivData}
 A data frame with the following columns: niv (the average net information
 value for each variable over all bootstrap samples), the penalty term, and
 the adjusted net information value.}
}
\description{
\code{niv} computes the net information value for each uplift predictor. This
can be a helpful exploratory tool to (preliminary) determine the predictive
power of each variable for uplift.
}
\details{
Given a binary response variable \eqn{y \in (0,1)}, the information value
(Siddiqi, 2006) from a predictor \eqn{x} is given by

\deqn{IV = \sum_{i=1}^{G} \left (P(x=i|y=1) - P(x=i|y=0) \right) \times WOE_i}

where \eqn{G} is the number of groups created from a numeric predictor or
levels from a categorical predictor, and \eqn{WOE_i = ln
(\frac{P(x=i|y=1)}{P(x=i|y=0)})}.

To avoid an undefined WOE, an adjustment factor A is used. Specifically,
\eqn{WOE_i = ln(\frac{(N(x=i|y=1)+A)/(N(y=1))}{(N(x=i|y=0)+A)/(N(y=0))})},
where \eqn{N} represents observation counts.

The net information value (NIV) proposed by Larsen (2009) is a natural
extension of the IV for the case of uplift. It is computed as

\deqn{NIV = \sum_{i=1}^{G}(P(x=i|y=1)^{T} \times P(x=i|y=0)^{C}  -
P(x=i|y=0)^{T} \times P(x=i|y=1)^{C}) \times NWOE_i} where \eqn{NWOE_i =
WOE_i^{T} - WOE_i^{C}}, and \eqn{T} and \eqn{C} refer to treatment and control
groups, respectively.

The adjusted net information value (ANIV) is computed as follows

\enumerate{ \item Draw B bootstrap samples from the training data and compute
the NIV for each variable in each sample. \item Compute the mean of the NIV
(\eqn{NIV_{mean}}) and sd of the NIV (\eqn{NIV_{sd}}) for each variable over
the \eqn{B} replications. \item The ANIV for a given variable is computed by
subtracting a penalty term from the mean NIV. Specifically, \eqn{ANIV =
NIV_{mean} - \frac{NIV_{sd}}{\sqrt{B}}}.}
}
\examples{
set.seed(1)
df <- sim_uplift(n = 1000, p = 20, response = "binary")
f <- create_uplift_formula(names(df)[-c(1:3)], "y", "T")
netInf <- niv(f, data = df, B=10, parallel = FALSE)
head(netInf$nivData)
}
\references{
Larsen, K. (2009). Net lift models. In: M2009 - 12th Annual SAS Data Mining
Conference.

Siddiqi, N. (2006). Credit Risk Scorecards: Developing and Implementing
Intelligent Credit Scoring. Wiley, Hoboken, NJ.
}
\seealso{
\code{\link{ggplot.niv}}.
}
\author{
Leo Guelman \email{leo.guelman@gmail.com}
}
