% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utree.R
\name{plot.utree}
\alias{plot.utree}
\title{Visualization of uplift trees.}
\usage{
\method{plot}{utree}(x, ...)
}
\arguments{
\item{x}{An object of class \code{utree}.}

\item{\dots}{Arguments passed to \code{partykit::plot.constparty}.}
}
\description{
plot method for class \code{utree}.
}
\examples{
set.seed(1)
df <- sim_uplift(n = 1000, p = 50, response = "binary")
form <- create_uplift_formula(x = names(df)[-c(1:3)], y = "y", trt = "T")
fit <- utree(form, data = df, maxdepth = 3)
plot(fit, main = "uplift tree", gp = gpar(cex = 0.5))
}
\author{
Leo Guelman \email{leo.guelman@gmail.com}
}
