% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uforest.R
\name{predict.uforest}
\alias{predict.uforest}
\title{Predict method for uforest fits.}
\usage{
\method{predict}{uforest}(object, newdata, ntrees = NULL, type = "uplift",
  agg.fun = "mean", ...)
}
\arguments{
\item{object}{A fitted object inheriting from \code{uforest}.}

\item{newdata}{A data frame to predict on.}

\item{ntrees}{Number of trees used in prediction. All fitted trees are
used by default.}

\item{type}{The type of predictions required. Only uplift predictions are
allowed.}

\item{agg.fun}{The function used to combine the predictions from the
individual trees. Must be functions that operate on a vector and produce a
single value, as \code{"mean"} and \code{"median"} do. It may be a
user-defined function.}

\item{\dots}{Additional arguments passed to \code{agg.fun}.}
}
\value{
A numeric vector of predictions.
}
\description{
Obtains predictions from a fitted \code{uforest} object.
}
\seealso{
\code{\link{uforest}}.
}
\author{
Leo Guelman \email{leo.guelman@gmail.com}
}
