% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uplift_glm.R
\name{predict.uplift_glm}
\alias{predict.uplift_glm}
\title{Predict method for uplift_glm fits.}
\usage{
\method{predict}{uplift_glm}(object, newdata = NULL, type = "uplift",
  na.action = na.omit, ...)
}
\arguments{
\item{object}{A fitted object inheriting from \code{uplift_glm}.}

\item{newdata}{An optional set of data to predict on. If \code{NULL}, then the
original data are used.}

\item{type}{The type of predictions required. Only \code{"uplift"} predictions
are allowed.}

\item{na.action}{The method for handling missing data.}

\item{\dots}{Additional arguments to be passed to other methods.}
}
\value{
A numeric vector of predictions for all methods, except when the model
 was fitted using \code{method = "glmnet"}, in which case the returned object
 is a matrix of predictions for the entire sequence of the penalty parameter
 used to create the model.
}
\description{
Obtains predictions from a fitted \code{uplift_glm} object.
}
\seealso{
\code{\link{uplift_glm}}
}
\author{
Leo Guelman \email{leo.guelman@gmail.com}
}
