% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utree.R
\name{predict.utree}
\alias{predict.utree}
\title{Predict method for utree fits.}
\usage{
\method{predict}{utree}(object, newdata, type = "uplift", ...)
}
\arguments{
\item{object}{A fitted object inheriting from \code{utree}.}

\item{newdata}{A data frame to predict on.}

\item{type}{The type of predictions required. The default is \code{"uplift"}
for uplift predictions Alternatively, \code{"node"}  returns an integer
vector of terminal node identifiers.}

\item{\dots}{Additional arguments passed to \code{partykit::predict.party}.}
}
\value{
A numeric vector of predictions.
}
\description{
Obtains predictions from a fitted \code{utree} object.
}
\seealso{
\code{\link{utree}}.
}
\author{
Leo Guelman \email{leo.guelman@gmail.com}
}
