% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_uplift.R
\name{sim_uplift}
\alias{sim_uplift}
\title{Uplift simulations.}
\usage{
sim_uplift(n = 1000, p = 20, rho = 0.2, beta.par = sqrt(6),
  sigma0 = sqrt(2), response = "gaussian")
}
\arguments{
\item{n}{The number of observations.}

\item{p}{The number of predictors.}

\item{rho}{The correlation between predictors.}

\item{beta.par}{Size of main effects. See details.}

\item{sigma0}{Multiplier of error term. See details.}

\item{response}{The type of response distribution. Possible values are
\code{"gaussian"} and \code{"binary"}.}
}
\value{
A data frame including the response variable (\code{y}), the treatment
 indicator (\code{T}), the "true" uplift effect (\code{trueUplift}), and the
 predictors (\code{X}).
}
\description{
Numerical simulations for uplift modeling, as described in Tian et al. (2014).
}
\details{
For the gaussian case, \code{sim_uplift} simulates data according to the
following specification:

\deqn{y = (\beta_0 + \sum_{j=1}^p \beta_{j}X_{j})^2 + (\gamma_0 + \sum_{j=1}^p
\gamma_{j}X_{j} + 0.8 X_1 X_2) T + \sigma_{0}\epsilon}

where the covariates \eqn{(X_{1}, \ldots, X_{p})} follow a mean zero
multivariate normal distribution with a compound symmetric variance-covariance
matrix, \eqn{(1-\rho)\mathbf{I}_{p} +\rho \mathbf{1}^{'}\mathbf{1}},
\eqn{\beta_0} = \code{beta.par}^-1, \eqn{\beta_j} =  (2 * \code{beta.par})^-1,
\eqn{\gamma_0 = 0.4}, \eqn{\gamma_j = (0.8, -0.8, 0.8, -0.8, 0, \ldots, 0)},
\eqn{T=[-1,1]} is the treatment indicator gerated with equal probability at
random, \eqn{\epsilon} is \eqn{N(0,1)}, and \eqn{\sigma_{0}} = \code{sigma0}.

For the binary case,

\deqn{y = I((\beta_0 + \sum_{j=1}^p \beta_{j}X_{j})^2 + (\gamma_0 +
\sum_{j=1}^p \gamma_{j}X_{j} + 0.8 X_1 X_2) T + \sigma_{0}\epsilon \ge 0)}

For further details, see Tian et al. (2014).
}
\examples{

set.seed(324)
df1 <- sim_uplift(p = 30, response = "binary")
str(df1)
df2 <- sim_uplift(n = 10000, p = 20)
str(df2)
}
\references{
Tian, L., Alizadeh, A., Gentles, A. and Tibshirani, R. (2014). "A simple
method for detecting interactions between a treatment and a large number of
covariates." Journal of the American Statistical Association, 109:508, pp.
1517--1532.
}
\author{
Leo Guelman \email{leo.guelman@gmail.com}
}
