% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trt}
\alias{trt}
\title{Mark treatment term.}
\usage{
trt(x)
}
\arguments{
\item{x}{The treatment indicator variable.}
}
\value{
Same as \code{x}.
}
\description{
An identity function, simply used to mark the treatment term in the model formula.
}
\author{
Leo Guelman \email{leo.guelman@gmail.com}
}
