% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uforest.R
\name{uforest_control}
\alias{uforest_control}
\title{Control for uplift random forest.}
\usage{
uforest_control(ntree = 100, minsplit = 40L, minbucket.t = 20L,
  minbucket.c = 20L, var.select.criterion = "pvalue",
  var.select.test = "asymptotic()", alpha = 0.05, bonferroni = FALSE,
  balance.sample = "undersample", split.criterion = "uplift",
  maxdepth = Inf, mtry = NULL, parallel = TRUE, nCore = NULL)
}
\arguments{
\item{ntree}{Number of uplift trees to fit.}

\item{minsplit}{The minimum number of observations in a node in order to be
considered for splitting.}

\item{minbucket.t}{The minimum number of treatment observations in any
terminal <leaf> node. The \code{treatLevel} can be used to determine the
treatment level of interest.}

\item{minbucket.c}{The minimum number of control observations in any terminal
<leaf> node.}

\item{var.select.criterion}{The criterion used to select the variable for
splitting. At the moment, only \code{"pvalue"} is accepted. The variable
with minimum pvalue is selected for splitting.}

\item{var.select.test}{The conditional null distribution of the test
statistic. This is passed to the \code{distribution} argument in
\code{coin::independence_test}. For example, for an approximative (Monte
Carlo) reference distribution with B Monte Carlo replicates, use
\code{approximate(B=999)}.}

\item{alpha}{The maximum acceptable pvalue required in order to make a split.}

\item{bonferroni}{Apply bonferroni adjustment to pvalue?}

\item{balance.sample}{The sampling method used to balance the treatment
variable. This attempts to have an equal representation of each treatment
before implementing the independence test described in
\code{var.select.test}. The options are \code{"undersample"} (default),
\code{"oversample"}, \code{"none"}. See the argument \code{sampling} in
\code{\link{mom}} for details.}

\item{split.criterion}{The split criteria used at each node of each tree;
Possible values are: \code{"uplift"} (default), \code{"kld"}
(Kullback-Leibler divergence), \code{"ed"} (Euclidean divergence),
\code{"l1d"} (L1-norm divergence). See details in Guelman et al. (2015).}

\item{maxdepth}{Maximum depth of the tree. The default \code{maxdepth = Inf}
means that no restrictions are applied to tree sizes.}

\item{mtry}{Number of input variables randomly sampled as candidates at each
node. The default is \eqn{\sqrt{p}}, where \eqn{p} represents the number of
covariates.}

\item{parallel}{If \code{TRUE}, computations are performed in parallel,
otherwise they are done sequentially.}

\item{nCore}{The number of cores used. Default is: number of available
cores-1.}
}
\value{
A list.
}
\description{
Various parameters that control aspects of the \code{uforest} fit.
}
\author{
Leo Guelman \email{leo.guelman@gmail.com}
}
